v1.3.5: drop-oldest recognition queue + integral-frame hardening
src/audio/recognizer.py: when the recognition queue is full, evict the
OLDEST chunk and admit the incoming one — freshest audio wins, matching
AudioCapture's block-queue policy. Previously the incoming chunk was
discarded and stale audio kept being processed first, delaying track-change
detection when Shazam lagged.

src/audio/capture.py + src/audio/chunking.py: fractional seconds in
config.yaml (e.g. chunk_seconds: 7.5) previously passed validation and
crashed mid-capture with a cryptic numpy TypeError. Capture now coerces
frame math to int; ChunkAssembler rejects fractional frame counts with a
clear ValueError (whole-valued floats from seconds arithmetic are accepted
and coerced).

+++ src/audio/capture.py
@@ -141,8 +141,12 @@ class AudioCapture:
         device_index = self._find_device_index()
         loop = asyncio.get_running_loop()
 
-        chunk_frames = self.chunk_seconds * self.sample_rate
-        hop_frames = (self.chunk_seconds - self.overlap_seconds) * self.sample_rate
+        # int() guards against fractional seconds in config.yaml (v1.3.5):
+        # float frame counts previously sailed through to numpy slicing and
+        # crashed mid-capture with a cryptic TypeError.  ChunkAssembler also
+        # validates integrality as a second line of defence.
+        chunk_frames = int(self.chunk_seconds * self.sample_rate)
+        hop_frames = int((self.chunk_seconds - self.overlap_seconds) * self.sample_rate)
         self._running = True
 
         log.info(
